# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go CRUD Backend API
  version: 1.0.0
  description: |
    API REST para gerenciamento de produtos.

    Contexto do backend:
    - Recurso principal: Product
    - Persistência atual: memória
    - Base path: /api/v1
    - Health check: /health
servers:
  - url: http://localhost:8080
    description: Ambiente local
  - url: http://127.0.0.1:8080
    description: Ambiente local alternativo
tags:
  - name: Health
    description: Verificação de disponibilidade da API
  - name: Products
    description: Operações de CRUD de produtos
paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Verifica se a API está online.
      operationId: healthCheck
      responses:
        '200':
          description: API online
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Listar produtos
      description: Retorna todos os produtos cadastrados, ordenados por ID.
      operationId: listProducts
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              example:
                - id: 1
                  name: Notebook
                  description: Notebook para desenvolvimento
                  category: Notebooks
                  image_url: https://source.unsplash.com/900x900/?laptop,developer
                  price: 4500.5
                  stock: 10
                  created_at: '2026-05-20T23:00:00Z'
                  updated_at: '2026-05-20T23:00:00Z'
    post:
      tags:
        - Products
      summary: Criar produto
      description: Cria um novo produto.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
            example:
              name: Notebook
              description: Notebook para desenvolvimento
              category: Notebooks
              image_url: https://source.unsplash.com/900x900/?laptop,developer
              price: 4500.5
              stock: 10
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: JSON inválido ou dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                invalidJson:
                  summary: JSON inválido
                  value:
                    message: json inválido
                validation:
                  summary: Dados inválidos
                  value:
                    message: dados inválidos
                    errors:
                      name: nome é obrigatório
                      price: preço não pode ser negativo
                      stock: estoque não pode ser negativo
  /api/v1/products/{id}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags:
        - Products
      summary: Buscar produto por ID
      description: Retorna um produto específico pelo ID.
      operationId: getProductById
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: id inválido
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: produto não encontrado
    put:
      tags:
        - Products
      summary: Atualizar produto
      description: Atualiza todos os dados de um produto existente.
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
            example:
              name: Notebook Pro
              description: Notebook atualizado
              category: Notebooks
              image_url: https://source.unsplash.com/900x900/?laptop,developer
              price: 5200
              stock: 8
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: ID, JSON ou dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: produto não encontrado
    delete:
      tags:
        - Products
      summary: Excluir produto
      description: Remove um produto pelo ID.
      operationId: deleteProduct
      responses:
        '204':
          description: Produto excluído com sucesso
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: id inválido
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: produto não encontrado
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: ID numérico do produto.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Notebook
        description:
          type: string
          example: Notebook para desenvolvimento
        category:
          type: string
          example: Notebooks
        image_url:
          type: string
          format: uri
          example: https://source.unsplash.com/900x900/?laptop,developer
        price:
          type: number
          format: double
          minimum: 0
          example: 4500.5
        stock:
          type: integer
          minimum: 0
          example: 10
        created_at:
          type: string
          format: date-time
          example: '2026-05-20T23:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-05-20T23:00:00Z'
    ProductInput:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
          minLength: 1
          example: Notebook
        description:
          type: string
          example: Notebook para desenvolvimento
        category:
          type: string
          example: Notebooks
        image_url:
          type: string
          format: uri
          example: https://source.unsplash.com/900x900/?laptop,developer
        price:
          type: number
          format: double
          minimum: 0
          example: 4500.5
        stock:
          type: integer
          minimum: 0
          example: 10
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: produto não encontrado
    ValidationErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: dados inválidos
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            name: nome é obrigatório